R5-G1-3 (clang-tidy narrowing tier, part 3): 39 qint64->double size displays made explicit
The adjudication set aside 40 findings the check reported as narrowing-to-int
but which are really qint64 -> double conversions: a byte count or an elapsed
time divided by a floating-point scale constant (kBytesPerMBf, kBytesPerGBf,
kMillisecondsPerSecondF, kNanosecondsPerMillisecond, ...) for a MB/GB/second
display. Casting the operand to int -- what a naive silencing would do -- would
truncate the value and corrupt the displayed size/duration, so those were left
for a value-preserving fix.

This makes each explicit the correct way: static_cast<double> on the integer
operand (the divisor a.size() in win32_mcp_fingerprint). The operand was already
promoted to double before the division, so the value is identical; the cast just
names the conversion so the narrowing check stops flagging it. onClearDownloadCaches
was refactored to hoist the scaled value into a local so the wrapped .arg calls
no longer push it past the 70-line function limit. One reported site had drifted
onto an unrelated line after the earlier edits and was correctly left alone.

clang-format + lizard clean, full Release build + ctest 225/225.

+++ src/core/advanced_uninstall_controller.cpp
@@ -880,7 +880,7 @@ void AdvancedUninstallController::onCleanupComplete(int succeeded,
         msg += QString(", %1 failed").arg(failed);
     }
     if (bytesRecovered > 0) {
-        const double mb = bytesRecovered / kBytesPerMBf;
+        const double mb = static_cast<double>(bytesRecovered) / kBytesPerMBf;
         msg += QString(" (%1 MB recovered)").arg(mb, 0, 'f', kRecoveredSizeDisplayPrecision);
     }
     Q_EMIT statusMessage(msg, kStatusTimeoutLongMs);
